+++ CHANGES.rst
@@ -7,6 +7,8 @@
 
 - Move support for large magnitude integer literals to Converter. [#1044]
 
+- Drop support for core/subclass_metadata-1.0.0 and core/constant-1.0.0 tags. [#1045]
+
 2.8.4 (unreleased)
 ------------------